+++ packages/esm-billing-app/translations/am.json
@@ -28,8 +28,6 @@
   "attributeRetiredReason": "ባህርይ ጡረታ የወጣበት ምክንያት",
   "balance": "ሚዛን",
   "benefits": "ጥቅሞች",
-  "Bill Deposit": "የሂሳብ ተቀማጭ",
-  "Bill Manager": "የሂሳብ አስተዳዳሪ",
   "billableExemptionAdministration": "የነጻነት አስተዳደር",
   "billableServicesError": "የክፍያ አገልግሎቶች ስህተት",
   "billAmount": "የሂሳብ መጠን",
@@ -59,7 +57,6 @@
   "cashier": "ገንዘብ ያዥ",
   "cashPoint": "የገንዘብ ነጥብ",
   "cashPointName": "የገንዘብ ነጥብ ስም",
-  "Charge Items": "የክፍያ ዕቃዎች",
   "ChargeableService": "Chargeable service",
   "chargeItem": "የክፍያ ዕቃዎች ሰንጠረዥ",
   "chargeItems": "የክፍያ ዕቃዎች",
@@ -76,7 +73,6 @@
   "claimExplanation": "የጥያቄ ማብራሪያ",
   "claimJustification": "የጥያቄ ቅድመ ምክንያት",
   "claims": "ጥያቄዎች",
-  "Claims Overview": "የጥያቄዎች አጠቃላይ እይታ",
   "claimsItems": "የጥያቄ ዕቃዎች",
   "claimsManagement": "የጥያቄዎች አስተዳደር",
   "claimSummary": "የጥያቄ ማጠቃለያ",
@@ -136,7 +132,6 @@
   "errorOnLine": "በመስመር ላይ ስህተት",
   "errorRetrievingHIESubscription": "HIE ደንበኝነት በማግኘት ላይ ስህተት",
   "exemptionCategory": "የነጻነት ምድብ",
-  "Exemptions": "ነጻነቶች",
   "facility": "ተቋም",
   "failedBillPayment": "የሂሳብ ክፍያ ወድቋል",
   "failure": "ጣልቃ ገብነቶችን በመጫን ላይ ስህተት",
@@ -152,7 +147,6 @@
   "incompletePaymentSubtitle": "እባክዎ ሁሉም የተመረጡ መስመር ዕቃዎች ሙሉ በሙሉ መከፈላቸውን ያረጋግጡ፣ የሚጠበቀው ጠቅላላ መጠን {{selectedLineItemsAmountDue}} ነው",
   "initiatePayment": "ክፍያ ያስጀምሩ",
   "inputSampleSchema": "የናሙና ስኬማ ግቤት",
-  "Insurance benefits": "የኢንሹራንስ ጥቅሞች",
   "insuranceScheme": "የኢንሹራንስ ዘዴ",
   "insurer": "ኢንሹራንስ አስተዳዳሪ",
   "interventionCode": "የጣልቃ ገብነት ኮድ",
@@ -209,7 +203,6 @@
   "overflowMenu": "የትርፍ ምናሌ",
   "overPayment": "ከመጠን በላይ ክፍያ",
   "overPaymentSubtitle": "የተከፈለው መጠን {{totalAmountTendered}} ለተመረጡ መስመር ዕቃዎች ከሚገባው መጠን {{selectedLineItemsAmountDue}} መብለጥ የለበትም",
-  "overview": "አጠቃላይ እይታ",
   "package": "ጥቅል",
   "patient": "ታካሚ",
   "patientBillingAlert": "የታካሚ ሂሳብ ማስጠንቀቂያ",
@@ -221,9 +214,6 @@
   "patientMissingSHAId": "ታካሚ የSHA መለያ ቁጥር ይጎድላል",
   "patientMissingSHANumber": "ታካሚ የSHA ቁጥር ይጎድላል፣ የSHA ማረጋገጫ መደረግ አይችልም፣ ታካሚው ወደ ምዝገባ ዴስክ እንዲሄድ ይመከራል",
   "patientName": "የታካሚ ስም",
-  "Payment History": "የክፍያ ታሪክ",
-  "Payment Modes": "የክፍያ ዘዴዎች",
-  "Payment Points": "የክፍያ ነጥቦች",
   "paymentHistory": "የክፍያ ታሪክ",
   "paymentMade": "ክፍያዎች ተከፍለዋል",
   "paymentMethod": "የክፍያ ዘዴ",
@@ -240,7 +230,6 @@
   "Phone Number": "ስልክ ቁጥር",
   "pleaseSelectItems": "ጥያቄ ለማቅረብ እባክዎ መስመር ዕቃዎችን ይምረጡ",
   "policyNumber": "የፖሊሲ ቁጥር",
-  "Pre-Auth Requests": "ቅድመ ፍቃድ ጥያቄዎች",
   "preauthRequest": "ቅድመ ፍቃድ ጥያቄዎች",
   "preAuthRequests": "ቅድመ ፍቃድ ጥያቄዎች",
   "preAuthRequets": "ቅድመ ፍቃድ ጥያቄዎች",
