+++ packages/esm-billing-app/translations/sw.json
@@ -28,8 +28,6 @@
   "attributeRetiredReason": "Attribute retired reason",
   "balance": "Balance",
   "benefits": "Benefits",
-  "Bill Deposit": "Bill Deposit",
-  "Bill Manager": "Bill Manager",
   "billableExemptionAdministration": "Exemption Administration",
   "billableServicesError": "Billable services error",
   "billAmount": "Bill Amount",
@@ -59,7 +57,6 @@
   "cashier": "Cashier",
   "cashPoint": "Cash Point",
   "cashPointName": "Cash Point Name",
-  "Charge Items": "Charge Items",
   "ChargeableService": "Chargeable service",
   "chargeItem": "Charge items table",
   "chargeItems": "Charge Items",
@@ -76,7 +73,6 @@
   "claimExplanation": "Claim Explanation",
   "claimJustification": "Claim Justification",
   "claims": "Claims",
-  "Claims Overview": "Claims Overview",
   "claimsItems": "Claims Items",
   "claimsManagement": "Claims management",
   "claimSummary": "Claim Summary",
@@ -136,7 +132,6 @@
   "errorOnLine": "Error on line",
   "errorRetrievingHIESubscription": "Error retrieving HIE subscription",
   "exemptionCategory": "Exemption category",
-  "Exemptions": "Exemptions",
   "facility": "Facility",
   "failedBillPayment": "Bill payment failed",
   "failure": "Error loading interventions",
@@ -152,7 +147,6 @@
   "incompletePaymentSubtitle": "Please ensure all selected line items are fully paid, Total amount expected is {{selectedLineItemsAmountDue}}",
   "initiatePayment": "Initiate Payment",
   "inputSampleSchema": "Input sample schema",
-  "Insurance benefits": "Insurance benefits",
   "insuranceScheme": "Insurance scheme",
   "insurer": "Insurer",
   "interventionCode": "Intervention Code",
@@ -209,7 +203,6 @@
   "overflowMenu": "Overflow menu",
   "overPayment": "Over payment",
   "overPaymentSubtitle": "Amount paid {{totalAmountTendered}} should not be greater than amount due {{selectedLineItemsAmountDue}} for selected line items",
-  "overview": "Overview",
   "package": "Packages",
   "patient": "Patient",
   "patientBillingAlert": "Patient Billing Alert",
@@ -221,9 +214,6 @@
   "patientMissingSHAId": "Patient missing SHA identification number",
   "patientMissingSHANumber": "Patient is missing SHA number, SHA validation cannot be done, Advise patient to visit registration desk",
   "patientName": "Patient Name",
-  "Payment History": "Payment History",
-  "Payment Modes": "Payment Modes",
-  "Payment Points": "Payment Points",
   "paymentHistory": "Payment History",
   "paymentMade": "Payments already made",
   "paymentMethod": "Payment method",
@@ -240,7 +230,6 @@
   "Phone Number": "Phone Number",
   "pleaseSelectItems": "Please select line items to raise a claim",
   "policyNumber": "Policy number",
-  "Pre-Auth Requests": "Pre-Auth Requests",
   "preauthRequest": "Preauth requests",
   "preAuthRequests": "Pre-auth requests",
   "preAuthRequets": "Pre-Auth Requests",
