+++ packages/esm-express-workflow-app/src/components/facility-dashboard/index.tsx
@@ -1,9 +1,8 @@
 import { getSyncLifecycle } from '@openmrs/esm-framework';
 
-import { moduleName } from '../../constants';
 import FacilityDashboard from './facility-dashboard.component';
+import { moduleName } from '../../constants';
 import { createLeftPanelLink } from '../../shared/dashboard-link/dashboard-link.component';
-import DashboardAuthorizationForm from './dashboard-authorization-form.modal';
 
 const options = {
   featureName: 'express-workflow',
@@ -18,5 +17,3 @@ export const facilityLeftPanelLink = getSyncLifecycle(
   }),
   options,
 );
-
-export const facilityDashboardAuthorizationModal = getSyncLifecycle(DashboardAuthorizationForm, options);
