+++ packages/esm-version-app/translations/am.json
@@ -1,13 +1,3 @@
 {
-  "Accounting": "የሂሳብ አያያዝ",
-  "Admissions": "የታካሚ ክፍል",
-  "Claims management": "የይገባኛል ጥያቄዎች አስተዳደር",
-  "clinicalConsultation": "ክሊኒካዊ ምክክር",
-  "Home": "መነሻ",
-  "Linkage services": "የማገናኛ አገልግሎቶች",
-  "Mortuary": "የሬሳ ክፍል",
-  "Patient services": "Patient services",
-  "Referrals": "ሪፈራሎች",
-  "Registration": "Registration",
   "systemInfo": "System Info"
 }
