+++ packages/esm-version-app/translations/en.json
@@ -1,13 +1,3 @@
 {
-  "Accounting": "Accounting",
-  "Admissions": "Admissions",
-  "Claims management": "Claims management",
-  "clinicalConsultation": "Clinical consultation",
-  "Home": "Home",
-  "Linkage services": "Linkage services",
-  "Mortuary": "Mortuary",
-  "Patient services": "Patient services",
-  "Referrals": "Referrals",
-  "Registration": "Registration",
   "systemInfo": "System Info"
 }
